backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
supabase==2.10.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-dotenv==1.0.1
google-generativeai==0.8.3
httpx==0.27.2
email-validator==2.2.0
python-multipart==0.0.17
openpyxl==3.1.5
